requirements.txt
Pillow
requests
cos-python-sdk-v5